requirements.txt
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.4.0
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
base58==2.1.1
bleach==4.1.0
blinker==1.4
brotlipy @ file:///D:/bld/brotlipy_1636012457296/work
cachetools==5.0.0
certifi==2020.6.20
cffi @ file:///D:/bld/cffi_1636046270630/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1643649795184/work
click==8.0.3
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
cryptography @ file:///D:/bld/cryptography_1639699517190/work
debugpy @ file:///D:/bld/debugpy_1636043518246/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1643888246732/work
gitdb==4.0.9
GitPython==3.1.26
ibm-cloud-sdk-core @ file:///home/conda/feedstock_root/build_artifacts/ibm-cloud-sdk-core_1642195623371/work
ibm-watson @ file:///home/conda/feedstock_root/build_artifacts/ibm-watson_1643277332202/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1642433548627/work
ipykernel @ file:///C:/ci_310/ipykernel_1642119594474/work/dist/ipykernel-6.4.1-py3-none-any.whl
ipython @ file:///D:/bld/ipython_1642634765247/work
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi @ file:///D:/bld/jedi_1637175265063/work
Jinja2==3.0.3
joblib @ file:///tmp/build/80754af9/joblib_1601912903842/work
jsonschema==4.4.0
jupyter-client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1642858610849/work
jupyter-core @ file:///D:/bld/jupyter_core_1636814417151/work
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
MarkupSafe==2.0.1
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1631080358261/work
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///C:/ci_310/mkl_random_1643050563308/work
mkl-service==2.4.0
nbclient==0.5.10
nbconvert==6.4.1
nbformat==5.1.3
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1638419302549/work
notebook==6.4.8
numpy @ file:///C:/ci/numpy_and_numpy_base_1639497294389/work
packaging==21.3
pandas==1.4.0
pandocfilters==1.5.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1638334955874/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==9.0.1
plotly @ file:///tmp/build/80754af9/plotly_1626182254919/work
prometheus-client==0.13.1
prompt-toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1643362612956/work
protobuf==3.19.4
pyarrow==6.0.1
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydeck==0.7.1
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1641580240686/work
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1638819640841/work
Pympler==1.0.1
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1643496850550/work
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks @ file:///D:/bld/pysocks_1635862594204/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
pywin32==303
pywinpty==2.0.1
pyzmq @ file:///D:/bld/pyzmq_1635877387650/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1641580202195/work
scikit-learn @ file:///C:/ci_310/scikit-learn_1642249118560/work
scipy @ file:///C:/ci_310/scipy_1642247684282/work
Send2Trash==1.8.0
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
smmap==5.0.0
streamlit==1.5.0
tenacity @ file:///C:/ci_310/tenacity_1642674931705/work
terminado==0.13.1
testpath==0.5.0
threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
toml==0.10.2
toolz==0.11.2
tornado @ file:///D:/bld/tornado_1635819761050/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1635260543454/work
tzdata==2021.5
tzlocal==4.1
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1641584929973/work
validators==0.18.2
watchdog==2.1.6
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
websocket-client @ file:///D:/bld/websocket-client_1636927238270/work
widgetsnbextension==3.5.2
win-inet-pton @ file:///D:/bld/win_inet_pton_1635832844911/work
